+++ python/tests/jet/test_task_based_cpu_contractor.py
@@ -4,28 +4,28 @@ import jet
 
 
 @pytest.mark.parametrize("dtype", ["complex64", "complex128"])
-class TestTaskBasedCpuContractor:
+class TestTaskBasedContractor:
     def test_default_constructor(self, dtype):
         """Tests that the default constructor is called."""
-        tbcc = jet.TaskBasedCpuContractor(dtype=dtype)
-        assert tbcc.name_to_tensor_map == {}
-        assert tbcc.name_to_parents_map == {}
-        assert tbcc.results == []
-        assert tbcc.reduction_result == jet.Tensor(dtype=dtype)
-        assert tbcc.flops == 0
-        assert tbcc.memory == 0
+        tbc = jet.TaskBasedContractor(dtype=dtype)
+        assert tbc.name_to_tensor_map == {}
+        assert tbc.name_to_parents_map == {}
+        assert tbc.results == []
+        assert tbc.reduction_result == jet.Tensor(dtype=dtype)
+        assert tbc.flops == 0
+        assert tbc.memory == 0
 
     def test_contract(self, dtype, tensor_network):
         """Tests that a tensor network can be contracted."""
         tn = tensor_network(dtype=dtype)
         path = jet.PathInfo(tn=tn, path=[[0, 1], [2, 3]])
-        tbcc = jet.TaskBasedCpuContractor(dtype=dtype)
+        tbc = jet.TaskBasedContractor(dtype=dtype)
 
-        assert tbcc.add_contraction_tasks(tn, path) == 0
-        assert tbcc.add_deletion_tasks() == 4
-        assert tbcc.add_reduction_task() == 1
+        assert tbc.add_contraction_tasks(tn, path) == 0
+        assert tbc.add_deletion_tasks() == 4
+        assert tbc.add_reduction_task() == 1
 
-        tbcc.contract()
+        tbc.contract()
 
         want_result = jet.Tensor(shape=[2], indices=["i"], data=[1, -1j], dtype=dtype)
         want_name_to_tensor_map = {
@@ -42,9 +42,9 @@ class TestTaskBasedCpuContractor:
             "3:ik": {"4:i:results[0]"},
         }
 
-        assert tbcc.name_to_tensor_map == want_name_to_tensor_map
-        assert tbcc.name_to_parents_map == want_name_to_parents_map
-        assert tbcc.results == [want_result]
-        assert tbcc.reduction_result == want_result
-        assert tbcc.flops == 18
-        assert tbcc.memory == 6
+        assert tbc.name_to_tensor_map == want_name_to_tensor_map
+        assert tbc.name_to_parents_map == want_name_to_parents_map
+        assert tbc.results == [want_result]
+        assert tbc.reduction_result == want_result
+        assert tbc.flops == 18
+        assert tbc.memory == 6